+++ instapy/util.py
@@ -2,6 +2,9 @@
 import time
 import datetime
 from math import ceil
+from math import radians
+from math import degrees as rad2deg
+from math import cos
 import random
 import re
 import regex
@@ -75,7 +78,7 @@ def is_private_profile(browser, logger, following=True):
 
     # double check with xpath that should work only when we not follwoing a
     # user
-    if is_private is True and not following:
+    if is_private and not following:
         logger.info("Is private account you're not following.")
         body_elem = browser.find_element_by_tag_name('body')
         is_private = body_elem.find_element_by_xpath(
@@ -185,6 +188,10 @@ def validate_username(browser,
             potency_ratio *= -1
             reverse_relationship = True
 
+        # division by zero is bad
+        followers_count = 1 if followers_count == 0 else followers_count
+        following_count = 1 if following_count == 0 else following_count
+
         if followers_count and following_count:
             relationship_ratio = (
                 float(followers_count) / float(following_count)
@@ -194,11 +201,11 @@ def validate_username(browser,
         logger.info(
             "User: '{}'  |> followers: {}  |> following: {}  |> relationship "
             "ratio: {}"
-            .format(username,
-                    followers_count if followers_count else 'unknown',
-                    following_count if following_count else 'unknown',
-                    truncate_float(relationship_ratio,
-                                   2) if relationship_ratio else 'unknown'))
+                .format(username,
+                        followers_count if followers_count else 'unknown',
+                        following_count if following_count else 'unknown',
+                        truncate_float(relationship_ratio,
+                                       2) if relationship_ratio else 'unknown'))
 
         if followers_count or following_count:
             if potency_ratio and not delimit_by_numbers:
@@ -206,10 +213,10 @@ def validate_username(browser,
                     inap_msg = (
                         "'{}' is not a {} with the relationship ratio of {}  "
                         "~skipping user\n"
-                        .format(username,
-                                "potential user" if not reverse_relationship
-                                else "massive follower",
-                                truncate_float(relationship_ratio, 2)))
+                            .format(username,
+                                    "potential user" if not reverse_relationship
+                                    else "massive follower",
+                                    truncate_float(relationship_ratio, 2)))
                     return False, inap_msg
 
             elif delimit_by_numbers:
@@ -219,7 +226,7 @@ def validate_username(browser,
                             inap_msg = (
                                 "User '{}'s followers count exceeds maximum "
                                 "limit  ~skipping user\n"
-                                .format(username))
+                                    .format(username))
                             return False, inap_msg
 
                     if min_followers:
@@ -227,7 +234,7 @@ def validate_username(browser,
                             inap_msg = (
                                 "User '{}'s followers count is less than "
                                 "minimum limit  ~skipping user\n"
-                                .format(username))
+                                    .format(username))
                             return False, inap_msg
 
                 if following_count:
@@ -236,7 +243,7 @@ def validate_username(browser,
                             inap_msg = (
                                 "User '{}'s following count exceeds maximum "
                                 "limit  ~skipping user\n"
-                                .format(username))
+                                    .format(username))
                             return False, inap_msg
 
                     if min_following:
@@ -244,7 +251,7 @@ def validate_username(browser,
                             inap_msg = (
                                 "User '{}'s following count is less than "
                                 "minimum limit  ~skipping user\n"
-                                .format(username))
+                                    .format(username))
                             return False, inap_msg
 
                 if potency_ratio:
@@ -253,11 +260,11 @@ def validate_username(browser,
                         inap_msg = (
                             "'{}' is not a {} with the relationship ratio of "
                             "{}  ~skipping user\n"
-                            .format(username,
-                                    "potential user" if not
-                                    reverse_relationship else "massive "
-                                                              "follower",
-                                    truncate_float(relationship_ratio, 2)))
+                                .format(username,
+                                        "potential user" if not
+                                        reverse_relationship else "massive "
+                                                                  "follower",
+                                        truncate_float(relationship_ratio, 2)))
                         return False, inap_msg
 
     if min_posts or max_posts or skip_private or skip_no_profile_pic or \
@@ -280,14 +287,14 @@ def validate_username(browser,
                 inap_msg = (
                     "Number of posts ({}) of '{}' exceeds the maximum limit "
                     "given {}\n"
-                    .format(number_of_posts, username, max_posts))
+                        .format(number_of_posts, username, max_posts))
                 return False, inap_msg
         if min_posts:
             if number_of_posts < min_posts:
                 inap_msg = (
                     "Number of posts ({}) of '{}' is less than the minimum "
                     "limit given {}\n"
-                    .format(number_of_posts, username, min_posts))
+                        .format(number_of_posts, username, min_posts))
                 return False, inap_msg
 
     """Skip users"""
@@ -312,7 +319,7 @@ def validate_username(browser,
             return False, "---> Sorry, couldn't get if user profile pic url\n"
         if (profile_pic in default_profile_pic_instagram or str(
                 profile_pic).find(
-                "11906329_960233084022564_1448528159_a.jpg") > 0) and (
+            "11906329_960233084022564_1448528159_a.jpg") > 0) and (
                 random.randint(0, 100) <= skip_no_profile_pic_percentage):
             return False, "{} has default instagram profile picture\n".format(
                 username)
@@ -525,8 +532,11 @@ def get_active_users(browser, username, posts, boundary, logger):
         "Getting active users who liked the latest {} posts:\n  {}".format(
             posts, message))
 
-    for count in range(1, posts + 1):
+    count = 1
+    checked_posts = 0
+    while count <= posts:
         try:
+            checked_posts += 1
             sleep_actual(2)
             try:
                 likers_count = browser.execute_script(
@@ -552,11 +562,39 @@ def get_active_users(browser, username, posts, boundary, logger):
                     likers_count = None
             try:
                 likes_button = browser.find_elements_by_xpath(
-                    "//button[contains(@class, '_8A5w5')]")[1]
-                click_element(browser, likes_button)
-                sleep_actual(5)
+                    "//button[contains(@class, '_8A5w5')]")
+                '''
+                    Len(likes_button) = 3 when user is followed and it's a post,
+                    but when user is followed and it's a video it is 2, so we avoid this
+                    and empty case.
+                '''
+                if len(likes_button) == 3 or len(likes_button) == 1:
+                    likes_button = likes_button[-1]
+                    click_element(browser, likes_button)
+                    sleep_actual(3)
+                else:
+                    raise NoSuchElementException
+
             except (IndexError, NoSuchElementException):
                 # Video have no likes button / no posts in page
+                logger.info("video found, try next post until we run out of posts")
+
+                # edge case of account having only videos,  or last post is a video.
+                if checked_posts >= total_posts:
+                    break
+                # if not reached posts(parameter) value, continue (but load next post)
+                if count != posts + 1:
+                    try:
+                        # click close button
+                        close_dialog_box(browser)
+
+                        # click next button
+                        next_button = browser.find_element_by_xpath(
+                            "//a[contains(@class, 'HBoOv')]"
+                            "[text()='Next']")
+                        click_element(browser, next_button)
+                    except Exception:
+                        logger.error('Unable to go to next profile post')
                 continue
 
             # get a reference to the 'Likes' dialog box
@@ -606,12 +644,7 @@ def get_active_users(browser, username, posts, boundary, logger):
                     sleep_actual(1.2)  # old value 5.6
                     sc_rolled += 1
 
-                """ Old method 1 """
-                # tmp_list = browser.find_elements_by_xpath(
-                #     "//a[contains(@class, 'FPmhX')]")
-
                 user_list = get_users_from_dialog(user_list, dialog)
-                # print("len(user_list): {}".format(len(user_list)))
 
                 # write & update records at Progress Tracker
                 if amount:
@@ -635,7 +668,7 @@ def get_active_users(browser, username, posts, boundary, logger):
                                 "Cor! Failed to get the desired amount of "
                                 "usernames but trying again.."
                                 "\t|> post:{}  |> attempt: {}\n"
-                                .format(posts, try_again + 1))
+                                    .format(posts, try_again + 1))
                             try_again += 1
                             too_many_requests += 1
                             scroll_it = True
@@ -652,26 +685,13 @@ def get_active_users(browser, username, posts, boundary, logger):
             logger.error("Ku-ku! There is an error searching active users"
                          "~\t{}\n\n".format(str(exc).encode("utf-8")))
 
-            """ Old method 2 """
-            # try:
-            #     tmp_list = browser.find_elements_by_xpath(
-            #         "//div[contains(@class, '_1xe_U')]/a")
-
-            #     if len(tmp_list) > 0:
-            #         logger.info(
-            #             "Post {}  |  Likers: found {}, catched {}".format(
-            #                 count, len(tmp_list), len(tmp_list)))
-
-            # except NoSuchElementException:
-            #     print("Ku-ku")
-
         for user in user_list:
             active_users.append(user)
 
         sleep_actual(1)
 
         # if not reached posts(parameter) value, continue
-        if count + 1 != posts + 1 and count != 0:
+        if count != posts + 1:
             try:
                 # click close button
                 close_dialog_box(browser)
@@ -684,6 +704,7 @@ def get_active_users(browser, username, posts, boundary, logger):
 
             except Exception:
                 logger.error('Unable to go to next profile post')
+        count += 1
 
     real_time = time.time()
     diff_in_minutes = int((real_time - start_time) / 60)
@@ -1112,8 +1133,8 @@ def highlight_print(username=None, message=None, priority=None, level=None,
         lower_char = None
 
     if (upper_char
-        and (show_logs
-             or priority == "workspace")):
+            and (show_logs
+                 or priority == "workspace")):
         print("\n{}".format(
             upper_char * int(ceil(output_len / len(upper_char)))))
 
@@ -1136,8 +1157,8 @@ def highlight_print(username=None, message=None, priority=None, level=None,
             print(message)
 
     if (lower_char
-        and (show_logs
-             or priority == "workspace")):
+            and (show_logs
+                 or priority == "workspace")):
         print("{}".format(
             lower_char * int(ceil(output_len / len(lower_char)))))
 
@@ -1254,7 +1275,7 @@ def ping_server(host, logger):
         if connectivity is False:
             logger.warning(
                 "Pinging the server again!\t~total attempts left: {}"
-                .format(ping_attempts))
+                    .format(ping_attempts))
             ping_attempts -= 1
             sleep(5)
 
@@ -1529,7 +1550,7 @@ def explicit_wait(browser, track, ec_params, logger, timeout=35, notify=True):
         if notify is True:
             logger.info(
                 "Timed out with failure while explicitly waiting until {}!\n"
-                .format(ec_name))
+                    .format(ec_name))
         return False
 
     return result
@@ -1934,7 +1955,7 @@ def get_users_from_dialog(old_data, dialog):
 def progress_tracker(current_value, highest_value, initial_time, logger):
     """ Provide a progress tracker to keep value updated until finishes """
     if (current_value is None or
-        highest_value is None or
+            highest_value is None or
             highest_value == 0):
         return
 
@@ -1959,15 +1980,15 @@ def progress_tracker(current_value, highest_value, initial_time, logger):
         tracker_line = "-----------------------------------"
         filled_index = int(progress_percent / 2.77)
         progress_container = (
-            "["
-            + tracker_line[:filled_index]
-            + "+"
-            + tracker_line[filled_index:]
-            + "]"
+                "["
+                + tracker_line[:filled_index]
+                + "+"
+                + tracker_line[filled_index:]
+                + "]"
         )
         progress_container = (
-            progress_container[:filled_index + 1].replace("-", "=")
-            + progress_container[filled_index + 1:]
+                progress_container[:filled_index + 1].replace("-", "=")
+                + progress_container[filled_index + 1:]
         )
 
         total_message = ("\r  {}/{} {}  {}%    "
@@ -2022,12 +2043,12 @@ def parse_cli_args():
     parser.add_argument(
         "-p", "--password", help="Password", type=str, metavar="123")
     parser.add_argument(
-        "-pd", "--page-delay", help="Implicit wait", type=int, metavar="25")
+        "-pd", "--page-delay", help="Implicit wait", type=int, metavar=25)
     parser.add_argument(
         "-pa", "--proxy-address", help="Proxy address",
         type=str, metavar="192.168.1.1")
     parser.add_argument(
-        "-pp", "--proxy-port", help="Proxy port", type=str, metavar="8080")
+        "-pp", "--proxy-port", help="Proxy port", type=int, metavar=8080)
 
     """ Auto-booleans: adding these flags ENABLE themselves automatically
     ```python quickstart.py --use-firefox```
@@ -2070,6 +2091,57 @@ def parse_cli_args():
     return args
 
 
+def get_cord_location(browser, location):
+    base_url = 'https://www.instagram.com/explore/locations/'
+    query_url = '{}{}{}'.format(base_url, location, "?__a=1")
+    browser.get(query_url)
+    json_text = browser.find_element_by_xpath('//body').text
+    data = json.loads(json_text)
+
+    lat = data['graphql']['location']['lat']
+    lon = data['graphql']['location']['lng']
+
+    return lat, lon
+
+
+def get_bounding_box(latitude_in_degrees, longitude_in_degrees, half_side_in_miles, logger):
+    if half_side_in_miles == 0:
+        logger.error("Check your Radius its lower then 0")
+        return {}
+    if latitude_in_degrees < -90.0 or latitude_in_degrees > 90.0:
+        logger.error("Check your latitude should be between -90/90")
+        return {}
+    if longitude_in_degrees < -180.0 or longitude_in_degrees > 180.0:
+        logger.error("Check your longtitude should be between -180/180")
+        return {}
+    half_side_in_km = half_side_in_miles * 1.609344
+    lat = radians(latitude_in_degrees)
+    lon = radians(longitude_in_degrees)
+
+    radius = 6371
+    # Radius of the parallel at given latitude
+    parallel_radius = radius * cos(lat)
+
+    lat_min = lat - half_side_in_km / radius
+    lat_max = lat + half_side_in_km / radius
+    lon_min = lon - half_side_in_km / parallel_radius
+    lon_max = lon + half_side_in_km / parallel_radius
+
+    lat_min = rad2deg(lat_min)
+    lon_min = rad2deg(lon_min)
+    lat_max = rad2deg(lat_max)
+    lon_max = rad2deg(lon_max)
+
+    bbox = {
+        "lat_min": lat_min,
+        "lat_max": lat_max,
+        "lon_min": lon_min,
+        "lon_max": lon_max
+    }
+
+    return bbox
+
+
 class CustomizedArgumentParser(ArgumentParser):
     """
      Subclass ArgumentParser in order to turn off
@@ -2092,4 +2164,3 @@ class CustomizedArgumentParser(ArgumentParser):
         will give the location of the 'argparse.py' file that have this method.
         """
         return []
-