+++ CHANGELOG.md
@@ -4,12 +4,62 @@ The **goal** of this file is explaining to the users of our project the notable 
 _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)_.
 
 
-## [0.1.3] - UNKNOWN
+## [0.3.0] - ADD DATE HERE
+### Added
+- Allowing `follow_by_tags` to interact with the user
+- Context manager to interaction calls in `like_by_tags` and `follow_likers`
+- Engagement pods feature 🙌 
+- Smart Hashtags based on locations `set_smart_location_hashtags`
+
+### Changed
+- Expose `threaded_session` of Instapy.end()
+
+### Fixed
+- `follow_likers` always fetches zero likers
+- Prevent division by zero in `validate_username`
+
+
+## [0.2.3] - 2019-03-01
+### Changed
+- Made Log in text checking more resilient 
+
+
+## [0.2.2] - 2019-02-21
+### Fixed
+- Chromedriver requirement now >= 2.44 instead of == 2.44
+
+
+## [0.2.1] - 2019-02-21
+### Fixed
+- xPath for Log In button
+
+
+## [0.2.0] - 2019-02-18
+### Added
+- Accept pending follow requests for private account
+- Feature to `follow_by_locations`
+- Proxy Authentication support for Firefox
+
+### Fixed
+- Only import instapy-chromedriver package when needed
+- Avoid user errors providing user names with caps
+- Fix get_active_users wrong behavior on videos
+- Bug in _CLI_ **argparsing** - `proxy_port` & `page_delay` are integers and not strings.
+- Selectors for finding comments and liking comments on posts
+- Temporarily turn off follow for `like_by_tags` interaction
+
+### Changed
+- Enable users interact by the comments of their own profiles.
+- Moved elements from docs folder to instapy-docs and instapy-research repositories
+
+
+## [0.1.3] - 2019-02-05
 ### Fixed
 - Fix "_Failed to load desired amount of users!_" issue.
 
 ### Added
 - Add _Progress Tracker_ to `get_users_through_dialog()` function.
+- Add Proxy Authentication for Firefox
 
 
 ## [0.1.2] - 2019-02-04